__pmResult: second round of changes
- change __pmAllocResult() arg, it is now int numpmid, not size_t need
- __pmAllocResult() may return NULL on failure, caller's responsibility
  to decide what to do
- shift all diagnostics in result.c from -Dpdubuf to -Dalloc
- add pmFreeResult(NULL) as special debugging case to dump pool state
  when -Dalloc in play
- fix fatal bug in pool linked-list maintenance below "free" routines

+++ src/include/pcp/libpcp.h
@@ -1362,7 +1362,7 @@ PCP_CALL extern int __pmSetRequestTimeout(double);
 PCP_CALL extern int __pmInProfile(pmInDom, const pmProfile *, int);
 
 /* __pmResult alloc/offsets */
-PCP_CALL extern __pmResult *__pmAllocResult(size_t);
+PCP_CALL extern __pmResult *__pmAllocResult(int);
 /*
  * __pmResult and pmResult are identical after the timestamp.
  * If we're exporting a pmResult from a __pmResult, this function
